refactor(forms): address feedback
Consolidate everything related to converting errors in one place
