+++ packages/forms/signals/compat/public_api.ts
@@ -12,6 +12,6 @@
  * Entry point for all public APIs of this package.
  */
 export * from './src/api/compat_form';
-export * from './src/api/compat_validation_error';
+export {CompatValidationError} from '../src/compat/validation_errors';
 export * from './src/api/di';
 export {SignalFormControl} from './src/signal_form_control/signal_form_control';
